+++ src/PHPStan/ConfiguredIntegrationStubFilesExtension.php
@@ -109,6 +109,10 @@ final class ConfiguredIntegrationStubFilesExtension implements StubFilesExtensio
             'majors' => [11, 12, 13],
             'files' => [__DIR__ . '/../../stubs/illuminate/support.stub'],
         ],
+        'symfony/stopwatch' => [
+            'majors' => [6, 7, 8],
+            'files' => [__DIR__ . '/../../stubs/symfony/stopwatch.stub'],
+        ],
     ];
 
     /**